+++ rollup.config.mjs
@@ -7,10 +7,11 @@ import fs from 'fs';
 import path from 'path';
 
 const COMPILED_RRWEB_PATH = 'build/rrweb-compiled.js';
+const BUNDLED_RRWEB_PATH = 'build/rrweb-bundled.js';
 
 const aliasRrweb = () => alias({
     entries: [
-        { find: '@mixpanel/rrweb', replacement: COMPILED_RRWEB_PATH },
+        { find: /rrweb-entrypoint(?:\.js)?$/, replacement: COMPILED_RRWEB_PATH },
     ]
 });
 
@@ -42,15 +43,25 @@ const MINIFY = process.env.MINIFY || process.env.FULL;
 
 // Main builds used to develop / iterate quickly
 const MAIN_BUILDS = [
-    // compile rrweb first to es5 with swc, we'll replace the import later on
     {
-        'input': '@mixpanel/rrweb',
+        'input': 'src/recorder/rrweb-entrypoint.js',
+        'output': [
+            {
+                file: BUNDLED_RRWEB_PATH,
+                format: 'es',
+            }
+        ],
+        plugins: [nodeResolve({browser: true})]
+    },
+    {
+        'input': BUNDLED_RRWEB_PATH,
         'output': [
             {
                 file: COMPILED_RRWEB_PATH,
+                format: 'es',
             }
         ],
-        plugins: [nodeResolve({browser: true}), swc({swc: {jsc: {target: 'es5'}}})]
+        plugins: [swc({swc: {jsc: {target: 'es5'}}})]
     },
 
     // IIFE recorder bundle that is loaded asynchronously