+++ src/autocapture/index.js
@@ -249,6 +249,11 @@ Autocapture.prototype._trackPageLeave = function(ev, currentUrl, currentScrollHe
         // User has navigated away already ending their impression.
         return;
     }
+
+    if (!this.getConfig(CONFIG_TRACK_PAGE_LEAVE) && !this.mp.is_recording_heatmap_data()) {
+        return;
+    }
+
     this.hasTrackedScrollSession = true;
     var viewportHeight = Math.max(document.documentElement.clientHeight, window.innerHeight || 0);
     var scrollPercentage = Math.round(Math.max(this.maxScrollViewDepth - viewportHeight, 0) / (currentScrollHeight - viewportHeight) * 100);
@@ -268,11 +273,8 @@ Autocapture.prototype._trackPageLeave = function(ev, currentUrl, currentScrollHe
         '$current_url': currentUrl || _.info.currentUrl(),
         '$viewportHeight': viewportHeight, // This is the fold line
         '$viewportWidth':  Math.max(document.documentElement.clientWidth, window.innerWidth || 0),
+        '$captured_for_heatmap': this.mp.is_recording_heatmap_data()
     }, DEFAULT_PROPS);
-
-    if (this.mp.is_recording_heatmap_data() && !this.getConfig(CONFIG_TRACK_PAGE_LEAVE)) {
-        props['$captured_for_heatmap'] = true;
-    }
 
     // Send with beacon transport to ensure event is sent before unload
     this.mp.track(MP_EV_PAGE_LEAVE, props, {transport: 'sendBeacon'});
@@ -447,7 +449,7 @@ Autocapture.prototype.initRageClickTracking = function() {
             return;
         }
 
-        if (this._rageClickTracker.isRageClick(ev['pageX'], ev['pageY'], currentRageClickConfig)) {
+        if (this._rageClickTracker.isRageClick(ev, currentRageClickConfig)) {
             this.trackDomEvent(ev, MP_EV_RAGE_CLICK);
         }
     }.bind(this);